Fix failed compilation

+++ food-truck-api/src/test/java/food/truck/api/recommendations/SearchServiceTest.java
@@ -226,13 +226,13 @@ class SearchServiceTest {
 
     private Review createHighReview() {
         Review review = new Review();
-        review.setRating((short) 5);
+        review.setRating(5.0);
         return review;
     }
 
     private Review createLowReview() {
         Review review = new Review();
-        review.setRating((short) 3);
+        review.setRating(3.0);
         return review;
     }
